add requested by Dawid and Lukasz changes

+++ src/main/java/pl/simpleascoding/tutoringplatform/dto/AdvertisementDTO.java
@@ -1,5 +1,7 @@
 package pl.simpleascoding.tutoringplatform.dto;
 
+import java.math.BigDecimal;
+
 public record AdvertisementDTO(String category, String authorUsername, String title, String description,
-                               int costPerHour) {
+                               BigDecimal costPerHour) {
 }
